Simplify the change by just checking immediately before writing.

+++ pkg/vendir/config/lock_config.go
@@ -5,7 +5,9 @@ package config
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
+	"io/fs"
 	"os"
 	"path/filepath"
 
@@ -52,14 +54,21 @@ func NewLockConfigFromBytes(bs []byte) (LockConfig, error) {
 }
 
 func (c LockConfig) WriteToFile(path string) error {
+	existingBytes, err := os.ReadFile(path)
+	if err != nil && !errors.Is(err, fs.ErrNotExist) {
+		return fmt.Errorf("Failed to check existing lock file: %w", err)
+	}
+
 	bs, err := c.AsBytes()
 	if err != nil {
 		return fmt.Errorf("Marshaling lock config: %s", err)
 	}
 
-	err = os.WriteFile(path, bs, 0600)
-	if err != nil {
-		return fmt.Errorf("Writing lock config: %s", err)
+	if bytes.Compare(existingBytes, bs) != 0 {
+		err = os.WriteFile(path, bs, 0600)
+		if err != nil {
+			return fmt.Errorf("Writing lock config: %s", err)
+		}
 	}
 
 	return nil
@@ -143,10 +152,4 @@ func (c LockConfig) MergeContents(path string, replaceCon LockDirectoryContents)
 		return fmt.Errorf("Expected to match exactly one directory, but did not match any")
 	}
 	return nil
-}
-
-func (c LockConfig) IsEqualTo(other LockConfig) bool {
-	myBytes, _ := c.AsBytes()
-	yourBytes, _ := other.AsBytes()
-	return bytes.Compare(myBytes, yourBytes) == 0
 }
